third_party/blink/web_tests/platform/generic/fast/xmlhttprequest/xmlhttprequest-responseXML-html-document-responsetype-quirks-expected.txt

This tests the XMLHttpRequest responseXML loading an HTML document in Quirks Mode with responseType "document".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseXML is non-null.
PASS xhr.responseXML instanceof Document is true
The two open <br> elements should be siblings of one another and not nested per HTML parser spec.
PASS children.length is 5
PASS children[0].tagName is "DIV"
PASS children[0].id is "description"
PASS children[1].tagName is "BR"
PASS children[2].tagName is "BR"
PASS children[3].tagName is "DIV"
PASS children[3].id is "console"
PASS children[4].tagName is "SCRIPT"
PASS children[1].id is not "break-tag"
PASS xhr.responseXML.compatMode is "BackCompat"
PASS xhr.responseXML.compatMode is "BackCompat"
PASS successfullyParsed is true

TEST COMPLETE
